+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## 1.0.0 (18 April 2021)
 
-- Initial commit, hello!
-- Supports 2 options: "threshold", "callback"
+* Initial commit, hello!
+* Supports 2 options: "threshold", "callback"